+++ OpenDreamShared/EngineUtils/Map.cs
@@ -1,14 +1,13 @@
 
 namespace OpenDreamShared.EngineUtils;
 
-public struct MapCoordinates(uint X, uint Y, uint Z) {
-    public uint X = X;
-    public uint Y = Y;
-    public uint Z = Z;
-
-    public Vector2i XY = new Vector2i((int)X, (int)Y);
+public struct MapCoordinates(int X, int Y, int Z) {
+    public int X = X;
+    public int Y = Y;
+    public int Z = Z;
+    public Vector2i XY = new Vector2i(X, Y);
 }
 
 public static class MapId {
-    public static readonly uint Nullspace = 0;
+    public static readonly int Nullspace = 0;
 }
